wip: deduplicate codebase deletion logic between kits:uninstall and functions:delete
